Refactor for library mode usage
